% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{check_is_xptr}
\alias{check_is_xptr}
\title{Test whether an object is an external pointer}
\usage{
check_is_xptr(s)
}
\arguments{
\item{x}{object to test}
}
\description{
Test whether an object is an external pointer
}
\keyword{internal}
