% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clandnstine-package.R
\docType{package}
\name{clandnstine}
\alias{clandnstine}
\alias{clandnstine-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/clandnstine}
\item BugReports: \url{https://gitlab.com/hrbrmstr/clandnstine/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
