% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doh-query.R
\name{doh_post}
\alias{doh_post}
\title{Make a DoH Request (POST/wireformat)}
\usage{
doh_post(name, type = "A",
  server_path = "https://dns.quad9.net/dns-query")
}
\arguments{
\item{name}{name to query for}

\item{type}{DNS query type (defaults to "\code{A}")}

\item{server_path}{full URL path to the DoH server quer endpoint (defaults to Quad9).}
}
\value{
\code{NULL} (if the query failed) or a \code{data.frame} (tibble)
}
\description{
Issue a \code{POST} wireformat query of type \code{type} for \code{name} to
the DoH endpoint specified at \code{server_path}.
}
\examples{
doh_post("rud.is", "A")
}
\references{
\url{https://tools.ietf.org/id/draft-ietf-doh-dns-over-https-05.html}
}
