% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from-wire.R
\name{from_wireformat}
\alias{from_wireformat}
\title{Convert a wireformat DNS records to an R list}
\usage{
from_wireformat(wf)
}
\arguments{
\item{wf}{DNS wireformat record as an R raw vector}
}
\value{
list
}
\description{
Convert a wireformat DNS records to an R list
}
\examples{
from_wireformat(
  as.raw(c(
    0x00, 0x00, 0x81, 0x80, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00,
    0x00, 0x00, 0x03, 0x77, 0x77, 0x77, 0x07, 0x65, 0x78, 0x61,
    0x6d, 0x70, 0x6c, 0x65, 0x03, 0x63, 0x6f, 0x6d, 0x00, 0x00,
    0x01, 0x00, 0x01, 0x03, 0x77, 0x77, 0x77, 0x07, 0x65, 0x78,
    0x61, 0x6d, 0x70, 0x6c, 0x65, 0x03, 0x63, 0x6f, 0x6d, 0x00,
    0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x80, 0x00, 0x04,
    0xC0, 0x00, 0x02, 0x01
  ))
)
}
