% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolver.R
\name{gdns_context}
\alias{gdns_context}
\title{Create a gdns DNS over TLS context and populate it with a resolver
for use in resolution functions}
\usage{
gdns_context(resolvers = "9.9.9.9")
}
\arguments{
\item{resolvers}{character vector of valid DNS over TLS resolvers;
Defaults to Quad9 (\code{9.9.9.9}).}
}
\description{
Create a gdns DNS over TLS context and populate it with a resolver
for use in resolution functions
}
\note{
\href{https://dnsprivacy.org/wiki/display/DP/DNS+Privacy+Test+Servers#DNSPrivacyTestServers-DoTservers}{DNS Privacy}
maintains a list of DNS over TLS servers.
}
\examples{
x <- gdns_context()
x <- gdns_context("1.1.1.1")
}
\seealso{
Other query functions: \code{\link{gdns_get_address}},
  \code{\link{gdns_query}}
}
\concept{query functions}
