% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdns_get_hosts}
\alias{gdns_get_hosts}
\title{Retreive the value of the localnames namespace}
\usage{
gdns_get_hosts(gctx)
}
\arguments{
\item{gctx}{gdns resolver context created with \code{\link[=gdns_resolver]{gdns_resolver()}}}
}
\description{
Retreive the value of the localnames namespace
}
