% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdns_get_resolvconf}
\alias{gdns_get_resolvconf}
\title{Retreive the value with which the context's upstream recursive servers and suffixes were initialized}
\usage{
gdns_get_resolvconf(gctx)
}
\arguments{
\item{gctx}{gdns resolver context created with \code{\link[=gdns_resolver]{gdns_resolver()}}}
}
\description{
Retreive the value with which the context's upstream recursive servers and suffixes were initialized
}
\seealso{
Other context functions: 
\code{\link{gdns_get_hosts}()},
\code{\link{gdns_get_resolution_type}()},
\code{\link{gdns_get_root_servers}()},
\code{\link{gdns_get_timeout}()},
\code{\link{gdns_get_tls_ca_file}()},
\code{\link{gdns_get_tls_ca_path}()},
\code{\link{gdns_get_transports}()},
\code{\link{gdns_set_hosts}()},
\code{\link{gdns_set_resolution_type}()},
\code{\link{gdns_set_round_robin_upstreams}()},
\code{\link{gdns_set_timeout}()},
\code{\link{gdns_set_tls_ca_file}()},
\code{\link{gdns_set_tls_ca_path}()},
\code{\link{gdns_update_resolvers}()}
}
\concept{context functions}
