% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdns_get_timeout}
\alias{gdns_get_timeout}
\title{Retreive the number of milliseconds to wait for request to return}
\usage{
gdns_get_timeout(gctx)
}
\arguments{
\item{gctx}{gdns resolver context created with \code{\link[=gdns_resolver]{gdns_resolver()}}}
}
\description{
Retreive the number of milliseconds to wait for request to return
}
