% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdns_get_tls_ca_path}
\alias{gdns_get_tls_ca_path}
\title{Retreive the value with which the context's upstream recursive servers and suffixes were initialized}
\usage{
gdns_get_tls_ca_path(gctx)
}
\arguments{
\item{gctx}{gdns resolver context created with \code{\link[=gdns_resolver]{gdns_resolver()}}}
}
\description{
Retreive the value with which the context's upstream recursive servers and suffixes were initialized
}
