% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdns_get_transports}
\alias{gdns_get_transports}
\title{Retreive what transports are used for DNS lookups.}
\usage{
gdns_get_transports(gctx)
}
\arguments{
\item{gctx}{gdns resolver context created with \code{\link[=gdns_resolver]{gdns_resolver()}}}
}
\description{
Retreive what transports are used for DNS lookups.
}
