% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolver.R
\name{gdns_query}
\alias{gdns_query}
\title{Arbitrary DNS queries}
\usage{
gdns_query(gcx, name, rr_type = "txt")
}
\description{
Perform any valid resource record inquiry for a given name. See \code{Details}.
}
\details{
This returns a fairly complex result object but that is the nature
of DNS queries. You're likely going to want what is in \code{$replies_tree$answer}
but the rest of the structure contains lovely metadata about the query and
remote query environment. There will eventually be "as data frame"-ish helpers
for this object.

Valid values for \code{rr_type}:
\itemize{
\item \code{a}
\item \code{a6}
\item \code{aaaa}
\item \code{afsdb}
\item \code{any}
\item \code{apl}
\item \code{atma}
\item \code{avc}
\item \code{axfr}
\item \code{caa}
\item \code{cdnskey}
\item \code{cds}
\item \code{cert}
\item \code{cname}
\item \code{csync}
\item \code{dhcid}
\item \code{dlv}
\item \code{dname}
\item \code{dnskey}
\item \code{doa}
\item \code{ds}
\item \code{eid}
\item \code{eui48}
\item \code{eui64}
\item \code{gid}
\item \code{gpos}
\item \code{hinfo}
\item \code{hip}
\item \code{ipseckey}
\item \code{isdn}
\item \code{ixfr}
\item \code{key}
\item \code{kx}
\item \code{l32}
\item \code{l64}
\item \code{loc}
\item \code{lp}
\item \code{maila}
\item \code{mailb}
\item \code{mb}
\item \code{md}
\item \code{mf}
\item \code{mg}
\item \code{minfo}
\item \code{mr}
\item \code{mx}
\item \code{naptr}
\item \code{nid}
\item \code{nimloc}
\item \code{ninfo}
\item \code{ns}
\item \code{nsap}
\item \code{nsap_ptr}
\item \code{nsec}
\item \code{nsec3}
\item \code{nsec3param}
\item \code{null}
\item \code{nxt}
\item \code{openpgpkey}
\item \code{opt}
\item \code{ptr}
\item \code{px}
\item \code{rkey}
\item \code{rp}
\item \code{rrsig}
\item \code{rt}
\item \code{sig}
\item \code{sink}
\item \code{smimea}
\item \code{soa}
\item \code{spf}
\item \code{srv}
\item \code{sshfp}
\item \code{ta}
\item \code{talink}
\item \code{tkey}
\item \code{tlsa}
\item \code{tsig}
\item \code{txt}
\item \code{uid}
\item \code{uinfo}
\item \code{unspec}
\item \code{uri}
\item \code{wks}
\item \code{x25}
\item \code{zonemd}
}
}
\examples{
x <- gdns_resolver()
gdns_query(x, "example.com")
}
\references{
\url{https://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml}
}
