% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdns_resolver}
\alias{gdns_resolver}
\title{Create a gdns DNS over TLS context and populate it with a resolver
for use in resolution functions}
\usage{
gdns_resolver(resolver = "9.9.9.9")
}
\arguments{
\item{resolver}{length 1 <chr> of a valid DNS over TLS resolver;
Defaults to Quad9 (\code{9.9.9.9}).}
}
\description{
Create a gdns DNS over TLS context and populate it with a resolver
for use in resolution functions
}
\examples{
x <- gdns_resolver()
}
