% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolver.R
\name{gdns_set_hosts}
\alias{gdns_set_hosts}
\title{Initialized the context's local names namespace with values from the given hosts file.}
\usage{
gdns_set_hosts(gctx, hosts_file)
}
\arguments{
\item{gctx}{gdns resolver context created with \code{\link[=gdns_resolver]{gdns_resolver()}}}

\item{hosts_file}{path to a valid \code{hosts} file (e.g. "\verb{/etc/hosts}). This value
will be \code{\link[=path.expand]{path.expand()}}ed.}
}
\description{
Initialized the context's local names namespace with values from the given hosts file.
}
\examples{
x <- gdns_context()
gdns_set_hosts(x, "/etc/hosts")
}
\seealso{
Other context functions: 
\code{\link{gdns_get_hosts}()},
\code{\link{gdns_get_resolution_type}()},
\code{\link{gdns_get_resolvconf}()},
\code{\link{gdns_get_root_servers}()},
\code{\link{gdns_get_timeout}()},
\code{\link{gdns_get_tls_ca_file}()},
\code{\link{gdns_get_tls_ca_path}()},
\code{\link{gdns_get_transports}()},
\code{\link{gdns_set_resolution_type}()},
\code{\link{gdns_set_round_robin_upstreams}()},
\code{\link{gdns_set_timeout}()},
\code{\link{gdns_set_tls_ca_file}()},
\code{\link{gdns_set_tls_ca_path}()},
\code{\link{gdns_update_resolvers}()}
}
\concept{context functions}
