% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-res-types.R
\name{gdns_set_resolution_type}
\alias{gdns_set_resolution_type}
\title{Specify whether DNS queries are performed with recursive lookups or as a stub resolver}
\usage{
gdns_set_resolution_type(gctx, res_type = c("stub", "recursive"))
}
\arguments{
\item{gctx}{gdns resolver context created with \code{\link[=gdns_resolver]{gdns_resolver()}}}

\item{res_type}{length 1 character vector of either "\code{stub}" or "\code{recursive}"}
}
\description{
Specify whether DNS queries are performed with recursive lookups or as a stub resolver
}
\examples{
x <- gdns_context()
x <- gdns_set_resolution_type(x, "stub")
}
