% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdns_set_round_robin_upstreams}
\alias{gdns_set_round_robin_upstreams}
\title{Set/unset context to round robin queries over the available upstreams
when resolving with the stub resolution type.}
\usage{
gdns_set_round_robin_upstreams(gctx, flag = TRUE)
}
\arguments{
\item{gctx}{gdns resolver context created with \code{\link[=gdns_resolver]{gdns_resolver()}}}

\item{flag}{if \code{TRUE} (the default) round robin queries when using more than one stub resolver,}
}
\description{
Set/unset context to round robin queries over the available upstreams
when resolving with the stub resolution type.
}
