% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdns_set_timeout}
\alias{gdns_set_timeout}
\title{Specify the number of milliseconds to wait for request to return}
\usage{
gdns_set_timeout(gctx, timeout)
}
\arguments{
\item{gctx}{gdns resolver context created with \code{\link[=gdns_resolver]{gdns_resolver()}}}

\item{timeout}{number of milliseconds (integer; i.e. not-fractional)}
}
\description{
Specify the number of milliseconds to wait for request to return
}
