% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdns_set_tls_ca_file}
\alias{gdns_set_tls_ca_file}
\title{Specify the file with CA certificates for verification purposes}
\usage{
gdns_set_tls_ca_file(gctx, ca_file)
}
\arguments{
\item{gctx}{gdns resolver context created with \code{\link[=gdns_resolver]{gdns_resolver()}}}

\item{ca_file}{file with Certificate Authority certificates}
}
\description{
Specify the file with CA certificates for verification purposes
}
