% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdns_set_tls_ca_path}
\alias{gdns_set_tls_ca_path}
\title{Specify where the location for CA certificates for verification purposes are located}
\usage{
gdns_set_tls_ca_path(gctx, ca_path)
}
\arguments{
\item{gctx}{gdns resolver context created with \code{\link[=gdns_resolver]{gdns_resolver()}}}

\item{ca_path}{directory with Certificate Authority certificates}
}
\description{
Specify where the location for CA certificates for verification purposes are located
}
