% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-transports.R
\name{gdns_set_transports}
\alias{gdns_set_transports}
\title{Specifies what transport(s) is/ar used for DNS lookups}
\usage{
gdns_set_transports(gctx, transports = c("tls", "udp", "tcp"))
}
\arguments{
\item{gctx}{gdns resolver context created with \code{\link[=gdns_resolver]{gdns_resolver()}}}

\item{transports}{character vector of any/all of "\code{udp}", "\code{tcp}" or "\code{tls}".
Order matters as it specifies that the library will use to try to
perform the lookups.}
}
\description{
Specifies what transport(s) is/ar used for DNS lookups
}
\examples{
x <- gdns_context()
x <- gdns_set_transports(x, "tls")
}
