% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_address}
\alias{get_address}
\title{Lookup IP Address(es) for a Given Host}
\usage{
get_address(host, resolver = "9.9.9.9")
}
\arguments{
\item{host}{host (length 1 <chr>)}

\item{resolver}{resolver address. defaults to Quad9; (length 1 <chr>)}
}
\description{
Lookup IP Address(es) for a Given Host
}
\examples{
get_address("r-project.org")
get_address("yahoo.com", "1.1.1.1")
get_address("microsoft.com", "8.8.8.8")
}
