% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{int_gdns_set_hosts}
\alias{int_gdns_set_hosts}
\title{Internal version of gdns_set_hosts()}
\usage{
int_gdns_set_hosts(gctx, hosts)
}
\description{
Internal version of gdns_set_hosts()
}
\keyword{internal}
