% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{int_gdns_set_resolution_type}
\alias{int_gdns_set_resolution_type}
\title{Internal version of gdns_set_resolution_type()}
\usage{
int_gdns_set_resolution_type(gctx, res_type)
}
\description{
Internal version of gdns_set_resolution_type()
}
\keyword{internal}
