% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{int_gdns_set_transports}
\alias{int_gdns_set_transports}
\title{Internal version of set_transports()}
\usage{
int_gdns_set_transports(gctx, trans)
}
\description{
Internal version of set_transports()
}
\keyword{internal}
