% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{int_gdns_update_resolvers}
\alias{int_gdns_update_resolvers}
\title{Internal version of gdns_update_resolvers}
\usage{
int_gdns_update_resolvers(gctx, resolvers)
}
\description{
Internal version of gdns_update_resolvers
}
\keyword{internal}
