% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_null_xptr_}
\alias{is_null_xptr_}
\title{Test whether an external pointer is null}
\usage{
is_null_xptr_(s)
}
\arguments{
\item{x}{object to test}
}
\description{
Test whether an external pointer is null
}
\keyword{internal}
