% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-context.R
\name{print.gctx}
\alias{print.gctx}
\title{Printer for gdns contexts}
\usage{
\method{print}{gctx}(x, ...)
}
\arguments{
\item{x}{gdns context object}

\item{...}{unused}
}
\description{
Printer for gdns contexts
}
\keyword{internal}
