% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolver.R
\name{print.gdns_response}
\alias{print.gdns_response}
\title{Printer for gdns_response objects}
\usage{
\method{print}{gdns_response}(x, ...)
}
\arguments{
\item{x}{a \code{gdns_response} object}

\item{...}{ignored}
}
\description{
Printer for gdns_response objects
}
\keyword{internal}
