% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doh-query.R
\name{tidy.gdns_doh_response}
\alias{tidy.gdns_doh_response}
\title{Tidy a DoH POST response}
\usage{
\method{tidy}{gdns_doh_response}(x, ...)
}
\arguments{
\item{x}{a DoH POST response}

\item{...}{unused}
}
\description{
Tidy a DoH POST response
}
