% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clandnstine-package.R
\docType{package}
\name{clandnstine}
\alias{clandnstine}
\alias{clandnstine-package}
\title{Perform Secure-by-default 'DNS' Queries}
\description{
Methods are provided to query 'Domain Name System' ('DNS') stub
and recursive resolvers for all 'DNS' resource record types using 'UDP',
TCP', and/or 'TLS' transport layers. 'DNS' query support is provided
by the 'getdns' (<getdnsapi.net>) C library.
}
\details{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/clandnstine}
\item BugReports: \url{https://gitlab.com/hrbrmstr/clandnstine/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
