% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clandnstine-package.R
\docType{package}
\name{clandnstine}
\alias{clandnstine}
\alias{clandnstine-package}
\title{Perform Secure-by-default 'DNS' Queries}
\description{
Something something 'DNS. Something something 'TLS'.
Something something 'getdns API/library'.
}
\details{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/clandnstine}
\item BugReports: \url{https://gitlab.com/hrbrmstr/clandnstine/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
