% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdns_get_address}
\alias{gdns_get_address}
\title{Resolve a host to an addrss}
\usage{
gdns_get_address(gctx, host)
}
\arguments{
\item{gctx}{gdns resolver context created with \code{\link[=gdns_resolver]{gdns_resolver()}}}

\item{host}{to lookup}
}
\description{
Resolve a host to an addrss
}
\examples{
x <- gdns_resolver()
gdns_get_address(x, "yahoo.com")
x \%>\% gdns_get_address("yahoo.com")
}
\seealso{
Other query functions: \code{\link{gdns_context}},
  \code{\link{gdns_query}}
}
\concept{query functions}
