% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdns_get_resolution_type}
\alias{gdns_get_resolution_type}
\title{Get the current resolution type setting}
\usage{
gdns_get_resolution_type(gctx)
}
\arguments{
\item{gctx}{gdns resolver context created with \code{\link[=gdns_resolver]{gdns_resolver()}}}
}
\description{
Get the current resolution type setting
}
\examples{
x <- gdns_context()
gdns_get_resolution_type(x)
}
