% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolver.R
\name{gdns_resolver}
\alias{gdns_resolver}
\title{Create a gdns DNS over TLS context and populate it with a resolver
for use in resolution functions}
\usage{
gdns_resolver(resolvers = "9.9.9.9")
}
\arguments{
\item{resolver}{character vector of valid DNS over TLS resolvers;
Defaults to Quad9 (\code{9.9.9.9}).}
}
\description{
Create a gdns DNS over TLS context and populate it with a resolver
for use in resolution functions
}
\examples{
x <- gdns_resolver()
x <- gdns_resolver("1.1.1.1")
}
