% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolver.R
\name{gdns_set_hosts}
\alias{gdns_set_hosts}
\title{Initialized the context's local names namespace with values from the given hosts file.}
\usage{
gdns_set_hosts(gctx, hosts_file)
}
\arguments{
\item{gctx}{gdns resolver context created with \code{\link[=gdns_resolver]{gdns_resolver()}}}

\item{hosts_file}{path to a valid \code{hosts} file (e.g. "\code{/etc/hosts}). This value
will be \code{\link[=path.expand]{path.expand()}}ed.}
}
\description{
Initialized the context's local names namespace with values from the given hosts file.
}
\examples{
x <- gdns_context()
x <- gdns_set_hosts(x, "/etc/hosts")
}
