% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-res-types.R
\name{gdns_set_resolution_type}
\alias{gdns_set_resolution_type}
\title{Specify whether DNS queries are performed with recursive lookups or as a stub resolver}
\usage{
gdns_set_resolution_type(gctx, res_type = c("stub", "recursive"))
}
\arguments{
\item{gctx}{gdns resolver context created with \code{\link[=gdns_resolver]{gdns_resolver()}}}

\item{res_type}{length 1 character vector of either "\code{stub}" or "\code{recursive}"}
}
\description{
Specify whether DNS queries are performed with recursive lookups or as a stub resolver
}
\examples{
x <- gdns_context()
x <- gdns_set_resolution_type(x, "stub")
}
\seealso{
Other context functions: \code{\link{gdns_get_hosts}},
  \code{\link{gdns_get_resolution_type}},
  \code{\link{gdns_get_resolvconf}},
  \code{\link{gdns_get_root_servers}},
  \code{\link{gdns_get_timeout}},
  \code{\link{gdns_get_tls_ca_file}},
  \code{\link{gdns_get_tls_ca_path}},
  \code{\link{gdns_get_transports}},
  \code{\link{gdns_set_hosts}},
  \code{\link{gdns_set_round_robin_upstreams}},
  \code{\link{gdns_set_timeout}},
  \code{\link{gdns_set_tls_ca_file}},
  \code{\link{gdns_set_tls_ca_path}},
  \code{\link{gdns_update_resolvers}}
}
\concept{context functions}
