% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdns_set_timeout}
\alias{gdns_set_timeout}
\title{Specify the number of milliseconds to wait for request to return}
\usage{
gdns_set_timeout(gctx, timeout)
}
\arguments{
\item{gctx}{gdns resolver context created with \code{\link[=gdns_resolver]{gdns_resolver()}}}

\item{timeout}{number of milliseconds (integer; i.e. not-fractional)}
}
\description{
Specify the number of milliseconds to wait for request to return
}
\seealso{
Other context functions: \code{\link{gdns_get_hosts}},
  \code{\link{gdns_get_resolution_type}},
  \code{\link{gdns_get_resolvconf}},
  \code{\link{gdns_get_root_servers}},
  \code{\link{gdns_get_timeout}},
  \code{\link{gdns_get_tls_ca_file}},
  \code{\link{gdns_get_tls_ca_path}},
  \code{\link{gdns_get_transports}},
  \code{\link{gdns_set_hosts}},
  \code{\link{gdns_set_resolution_type}},
  \code{\link{gdns_set_round_robin_upstreams}},
  \code{\link{gdns_set_tls_ca_file}},
  \code{\link{gdns_set_tls_ca_path}},
  \code{\link{gdns_update_resolvers}}
}
\concept{context functions}
