% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolver.R
\name{gdns_update_resolvers}
\alias{gdns_update_resolvers}
\title{Changes the list of resolvers in an already created context for use in resolution functions}
\usage{
gdns_update_resolvers(gctx, resolvers)
}
\arguments{
\item{gctx}{gdns resolver context created with \code{\link[=gdns_resolver]{gdns_resolver()}}}

\item{resolvers}{character vector of valid DNS over TLS resolvers}
}
\description{
Changes the list of resolvers in an already created context for use in resolution functions
}
\note{
\href{https://dnsprivacy.org/wiki/display/DP/DNS+Privacy+Test+Servers#DNSPrivacyTestServers-DoTservers}{DNS Privacy}
maintains a list of DNS over TLS servers.
}
\examples{
x <- gdns_context()
x <- gdns_update_resolvers("1.1.1.1")
}
